PELE-438 Renamed constraints module to alpha_constraints
